services/api_fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
python-dotenv==1.0.1
pymongo[srv]==4.7.3
sentence-transformers==3.4.1
numpy>=1.26.0
newsapi==0.1.1
requests>=2.31.0
openai>=1.0.0
email-validator>=2.0.0
